server/requirements_2021-8-4.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
joblib==1.0.1
Markdown==3.3.4
numpy==1.21.1
psycopg2==2.9.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.2.0
